adds a new feature which computes the percentile latency(platency).
The goal is to exclude anomalies so as to give a better average.  Otherwise, the arithmetic average is inaccurate when very high or low latency requests present. The actual implementation will subject to change where the goal is determined.

The percentile is implemented as throw away 1/2 requests of total requests from the beginning and the end of a sorted request's latency array.
So the current implementation is simlar to the scoring rules of diving.